App/requirements.txt
pandas==1.3.3
numpy==1.21.2
sentence-transformers==2.2.0
faiss-cpu==1.7.1
openai==0.10.2
scikit-learn==0.24.2
requests==2.26.0
plotly==5.3.1
ipywidgets==7.6.5
streamlit==1.0.0
